:sparkles: feat: add local data classes
The LocalData class, as well as its implementations, will be used to store instances of entities, by id, that have already been loaded previously, without the need to instantiate them again.

Note: The entity service will still perform its commands, and in the end, it will result in a new entity (if an entity with the same id has not been generated previously) or in the old class with its updated data.